+++ CHANGELOG.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - WinUI: Added `IWindowManager.ShowContentDialog(viewModel, ownerViewModel, context)`
 
 ### Changed
-- `IWindowManager.ShowDialog(viewModel, ownerViewModel, context)` throws when owner Window cannot be determined
+- `IWindowManager` methods throw when owner window cannot be determined
 
 ### Removed
 - WPF: Removed `IWindowManager.ShowPopup(viewModel, context)`